requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.2
backoff==2.2.1
blinker==1.9.0
boto3==1.39.11
botocore==1.39.11
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
dacite==1.6.0
dagshub==0.6.2
dagshub-annotation-converter==0.1.11
databricks-sdk==0.59.0
dataclasses-json==0.6.7
docker==7.1.0
dotenv==0.9.9
fastapi==0.116.1
Flask==3.1.1
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
gql==3.5.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
mlflow==3.1.4
mlflow-skinny==3.1.4
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.1
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
packaging==25.0
pandas==2.3.1
pathvalidate==3.3.1
pillow==11.3.0
pipx==1.7.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.13.1
scikit-learn==1.7.1
scipy==1.16.0
semver==3.0.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.47.2
tenacity==9.1.2
threadpoolctl==3.6.0
treelib==1.8.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
userpath==1.9.2
uvicorn==0.35.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
